cmf fixes
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@2965 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/utils_mod.f90
@@ -3846,7 +3846,7 @@ END SUBROUTINE GAUSSJ
 
 
   subroutine  convertToFnu(nuarray, fArray, n)
-    real :: nuArray(:), fArray(:), nu
+    real(double) :: nuArray(:), fArray(:), nu
     real, allocatable :: nuTemp(:), fTemp(:)
     integer :: i, n
